+++ src/ui/menu/room/privacy_settings_popup.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <defs/geode.hpp>
+
+class PrivacySettingsPopup : public geode::Popup<> {
+public:
+    static PrivacySettingsPopup* create();
+
+protected:
+    bool setup() override;
+
+    void sendPacket();
+};